+++ widgets/ForwardWidgets-huangxd/widgets/danmu_auto.js
@@ -163,26 +163,56 @@ WidgetMetadata = {
       name: "danmu_api_1",
       title: "danmu_api_1 (前面匹配不到的可以试试弹幕API，比如一些韩剧/美剧)",
       type: "input",
+      placeholders: [
+        {
+          title: "示例",
+          value: "https://{domain}/api/{token}",
+        },
+      ],
     },
     {
       name: "danmu_api_2",
       title: "danmu_api_2 (前面匹配不到的可以试试弹幕API，比如一些韩剧/美剧)",
       type: "input",
+      placeholders: [
+        {
+          title: "示例",
+          value: "https://{domain}/api/{token}",
+        },
+      ],
     },
     {
       name: "danmu_api_3",
       title: "danmu_api_3 (前面匹配不到的可以试试弹幕API，比如一些韩剧/美剧)",
       type: "input",
+      placeholders: [
+        {
+          title: "示例",
+          value: "https://{domain}/api/{token}",
+        },
+      ],
     },
     {
       name: "danmu_api_4",
       title: "danmu_api_4 (前面匹配不到的可以试试弹幕API，比如一些韩剧/美剧)",
       type: "input",
+      placeholders: [
+        {
+          title: "示例",
+          value: "https://{domain}/api/{token}",
+        },
+      ],
     },
     {
       name: "danmu_api_5",
       title: "danmu_api_5 (前面匹配不到的可以试试弹幕API，比如一些韩剧/美剧)",
       type: "input",
+      placeholders: [
+        {
+          title: "示例",
+          value: "https://{domain}/api/{token}",
+        },
+      ],
     },
     {
       name: "api_priority",