+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-25 03:07:50
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 95dc7b7bbf01d96812eacd59ddb8489ef3777580
+
+## [ForwardWidgets-2kuai] - 2025-08-25 03:07:50
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: ae5e3dce37a5e06e77bbb58c05ebfe40befcc5ca
+
+## [ForwardWidgets-pack1r] - 2025-08-25 03:07:50
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 04847114cf9daf01542c32a93b9ef3597e7e35ca
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-24 03:11:38
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main